增加用户更新接口，更改登录流程

+++ src/main/java/com/changyi/fi/component/invoice/InvoiceResource.java
@@ -11,6 +11,8 @@ import com.changyi.fi.core.LogUtil;
 import com.changyi.fi.core.response.NormalResponse;
 import com.changyi.fi.core.token.Token;
 import com.changyi.fi.exception.AuthenticationFailedException;
+import com.changyi.fi.exception.NullRequestException;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -34,8 +36,8 @@ public class InvoiceResource {
     @Produces("application/json")
     @Secured
     public Response listInvoices(@HeaderParam(Token.KEY) String token) {
-        LogUtil.info(this.getClass(), "Enter invoices endpoint");
         try {
+            LogUtil.info(this.getClass(), "Enter invoices endpoint");
             InvoicesResponse response = invoiceService.listInvoice(Token.touch(token).getOpenId());
             LogUtil.info(this.getClass(), "Complete invoices endpoint handle");
             LogUtil.debug(this.getClass(), "Response: {} ", response.build());
@@ -52,9 +54,12 @@ public class InvoiceResource {
     @Produces("application/json")
     @Secured
     public Response updateInvoice(@HeaderParam(Token.KEY) String token, @RequestParam String request) {
-        LogUtil.info(this.getClass(), "Enter updateInvoice endpoint");
-        LogUtil.debug(this.getClass(), "Request: {} ", request);
         try {
+            LogUtil.info(this.getClass(), "Enter updateInvoice endpoint");
+            LogUtil.debug(this.getClass(), "Request: {} ", request);
+            if (StringUtils.isEmpty(request)) {
+                throw new NullRequestException("Request is required");
+            }
             PutInvoiceRequest req = new Payload(request).as(PutInvoiceRequest.class);
             invoiceService.updateInvoice(req, Token.touch(token).getOpenId());
             LogUtil.info(this.getClass(), "Complete addInvoice service call");
@@ -71,8 +76,11 @@ public class InvoiceResource {
     @Produces("application/json")
     @Secured
     public Response deleteInvoice(@HeaderParam(Token.KEY) String token, @PathParam("id") String id) {
-        LogUtil.info(this.getClass(), "Enter deleteInvoice endpoint for id: " + id);
         try {
+            LogUtil.info(this.getClass(), "Enter deleteInvoice endpoint for id: " + id);
+            if (StringUtils.isEmpty(id)) {
+                throw new NullRequestException("Id is required");
+            }
             invoiceService.deleteInvoice(Token.touch(token).getOpenId(), id);
             LogUtil.info(this.getClass(), "Complete deleteInvoice endpoint handle");
             return Response.status(Response.Status.OK).entity(new NormalResponse().build()).build();
@@ -88,8 +96,11 @@ public class InvoiceResource {
     @Produces("application/json")
     @Secured
     public Response getInvoice(@HeaderParam(Token.KEY) String token, @PathParam("id") String id) {
-        LogUtil.info(this.getClass(), "Enter getInvoice endpoint for id: " + id);
         try {
+            LogUtil.info(this.getClass(), "Enter getInvoice endpoint for id: " + id);
+            if (StringUtils.isEmpty(id)) {
+                throw new NullRequestException("Id is required");
+            }
             GetInvoiceResponse response = invoiceService.getInvoice(Token.touch(token).getOpenId(), id);
             LogUtil.info(this.getClass(), "Complete getInvoice endpoint handle");
             LogUtil.debug(this.getClass(), "Response: {} ", response.build());
